v8-langgraph-agent-advanced-prompt/server/pyproject.toml
[tool.poetry]
name = "ai-agent-poc"
version = "0.3.0"
description = "A proof of concept project using FastAPI"
authors = ["Your Name <viktor.vasylkovskyi@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = ">=0.109.0"
starlette = ">=0.38.6"
uvicorn = ">=0.26.0"
python-socketio = "^5.11.4"
ipython = "^8.30.0"
requests = "^2.32.3"
colorama = "^0.4.6"
python-dotenv = "^1.0.1"
openai = "^1.59.4"
wikipedia = "^1.4.0"
anonllm = "^0.1.10"
tiktoken = "^0.8.0"
google-cloud-aiplatform = "^1.76.0"
google-auth = "^2.37.0"
langgraph = "^0.2.61"
langchain-core = "^0.3.29"
langchain-google-vertexai = "^2.0.10"
langchain-community = "^0.3.14"
langchain = "^0.3.14"
chromadb = "^0.6.2"
langsmith = "^0.2.10"
langchain-openai = "^0.3.0"
pandas = "^2.2.3"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
pylint = "^3.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
